+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Features
   * Allow renaming JNI_OnLoad via define.
   * Enable CMake modules for use of Gluecodium in multiple targets.
+### Bug fixes:
+  * Fixed Swift compilation issue for lambdas with parameters of class or interface types.
 
 ## 6.0.1
 Release date: 2020-01-16